moved projectslist to a different folder
